Code cleanup
Code cleanup including:
- removing dead-code
- improving consitency across the files
- adding enhanced description of code
