+++ language_mod.py
@@ -1,8 +1,11 @@
+''' This file contains callback functions 
+to modify dashboard labels depending on 
+a language selected by a user'''
+
 from dash import Input, Output, callback
-import app_layout as al
 
 def lang_callbacks(app):
-    ### Chart title language change
+    ### Dashboard title (language change)
     @app.callback(
         Output('title_main', 'children'),
         Input('lang_sel', 'value')        
@@ -12,7 +15,7 @@ def lang_callbacks(app):
             return "Mortgage Calculation & Analysis"
         else:
             return "Symulacja Kosztów Kredytu"
-    #### Inputs tab langauage change
+    #### Inputs tab (langauage change)
     @app.callback(
         Output('title_input', 'children'),
         Output('loan_amount_label', 'children'),
@@ -32,7 +35,7 @@ def lang_callbacks(app):
             return "Input Data", "Mortgage amount", "Mortgage duration (years + months)", "Provide number of years", "Provide number of months", "Total number of installments", "Installment type", [{'label': 'Fixed', 'value': 'fixed'}, {'label': 'Descending', 'value': 'desc'}], "Mortgage interest (bank + WIBOR)", "Bank %", "WIBOR %"
         else:
             return "Dane", "Kwota kredytu", "Czas spłaty (lata i miesiące)", "Podaj liczbę lat", "Podaj liczbę miesięcy", "Całkowita ilość rat", "Rodzaj raty", [{'label': 'Stała', 'value': 'fixed'}, {'label': 'Malejąca', 'value': 'desc'}], "Oprocentowanie (marża banku + WIBOR)", "Marża %", "WIBOR %"
-    #### Outputs tab langauage change
+    #### Outputs tab - menu (langauage change)
     @app.callback(
         Output('title_output', 'children'),
         Output('overview_tab_label', 'label'),
@@ -46,7 +49,7 @@ def lang_callbacks(app):
             return "Details of mortgage simulation", "Overview", "Payment over time", "Amortization schedule", "WIBOR effect"
         else:
             return "Wyniki symulacji kredytu", "Podsumowanie", "Struktura płatności", "Harmonogram spłaty", "Wpływ zmian WIBOR"
-    #### Output - Overview tab language change
+    #### Output - Overview tab (language change)
     @app.callback(
         Output('kpi_label', 'children'),
         Output('kpi_note', 'children'),
@@ -57,8 +60,8 @@ def lang_callbacks(app):
         if lang == 1:
             return "Your Installment per Month:", "Please note that for descending installment type, above amount is for 1st installment only.", "Payment Breakdown",
         else:
-            return "Wysokość Twojej miesiącznej raty:", "Uwaga! Dla raty stałej powyższa kwota dotyczy tylko pierwszej raty (kolejne maleją)", "Składowe Wpłat",
-    #### Output - Payment over time tab language change
+            return "Wysokość Twojej miesięcznej raty:", "Uwaga! Dla raty stałej powyższa kwota dotyczy tylko pierwszej raty (kolejne maleją)", "Składowe Wpłat",
+    #### Output - Payment over time tab (language change)
     @app.callback(
         Output('pay_label', 'children'),
         Output('radioitems-payment_scatter', 'options'),
@@ -69,7 +72,7 @@ def lang_callbacks(app):
             return "Cumulative mortgage payments over time - view by:", [{"label": "Month", "value": 1}, {"label": "Year", "value": 2},]
         else:
             return "Suma wpłat na spłatę kredytu w czasie jego trwania - widok:", [{"label": "Miesięczny", "value": 1}, {"label": "Roczny", "value": 2},]
-    #### Output - Amortization schedule language change
+    #### Output - Amortization schedule tab (language change)
     @app.callback(
         Output('amort_label', 'children'),
         Output('radioitems-payment_table', 'options'),
@@ -80,3 +83,13 @@ def lang_callbacks(app):
             return "Amortization Schedule - view by:", [{"label": "Month", "value": 1}, {"label": "Year", "value": 2},]
         else:
             return "Harmonogram spłaty kredytu - widok:", [{"label": "Miesięczny", "value": 1}, {"label": "Roczny", "value": 2},]
+    # #### Output - Amortization schedule - dataframes language change
+    # @app.callback(
+    #     Output('installments_df', 'data'),
+    #     Output('installments_df_yr', 'data'),
+    #     Input('installments_df', 'data'),
+    #     Input('installments_df_yr', 'data'),
+    #     Input('lang_sel', 'value'),
+    # )
+    # def mod_df_headers_lang(data_mo, data_yr, lang):
+    #     return data_mo, data_yr
